refactor: relocate `shadcn/ui` components to `components/ui/`
