+++ src/components/ui/form.tsx
@@ -10,7 +10,7 @@ import {
 	useFormContext,
 } from "react-hook-form";
 
-import { Label } from "@/components/shadcn_ui/label";
+import { Label } from "@/components/ui/label";
 import { cn } from "@/lib/utils";
 
 const Form = FormProvider;
