+++ src/components/ui/use-toast.ts
@@ -1,10 +1,7 @@
 // Inspired by react-hot-toast library
 import * as React from "react";
 
-import type {
-	ToastActionElement,
-	ToastProps,
-} from "@/components/shadcn_ui/toast";
+import type { ToastActionElement, ToastProps } from "@/components/ui/toast";
 
 const TOAST_LIMIT = 1;
 const TOAST_REMOVE_DELAY = 1000000;
