+++ src/features/contact_me/components/MessageMe.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button } from "@/components/shadcn_ui/button";
+import { Button } from "@/components/ui/button";
 import {
 	Dialog,
 	DialogContent,
@@ -8,7 +8,7 @@ import {
 	DialogHeader,
 	DialogTitle,
 	DialogTrigger,
-} from "@/components/shadcn_ui/dialog";
+} from "@/components/ui/dialog";
 import {
 	Drawer,
 	DrawerContent,
@@ -16,7 +16,7 @@ import {
 	DrawerHeader,
 	DrawerTitle,
 	DrawerTrigger,
-} from "@/components/shadcn_ui/drawer";
+} from "@/components/ui/drawer";
 import MessageMeForm from "@/features/contact_me/components/MessageMeForm";
 import { Breakpoints, useMediaQuery } from "@/hooks/use-media-query";
 import { cn } from "@/lib/utils";
